Move unit tests to correct folder
The setup of the integration tests is slow and all unit tests should be
moved to the propper subvodler in test/unit.
